Component development. Page template and topnav element created.
