+++ core/src/main/java/trainingproject/core/filters/LoggingFilter.java
@@ -1,0 +1,67 @@
+/*
+ *  Copyright 2015 Adobe Systems Incorporated
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package trainingproject.core.filters;
+
+import java.io.IOException;
+import javax.servlet.Filter;
+import javax.servlet.FilterChain;
+import javax.servlet.FilterConfig;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
+
+import org.apache.sling.api.SlingHttpServletRequest;
+import org.apache.sling.engine.EngineConstants;
+import org.osgi.framework.Constants;
+import org.osgi.service.component.annotations.Component;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * Simple servlet filter component that logs incoming requests.
+ */
+@Component(service = Filter.class,
+           property = {
+                   Constants.SERVICE_DESCRIPTION + "=Demo to filter incoming requests",
+                   EngineConstants.SLING_FILTER_SCOPE + "=" + EngineConstants.FILTER_SCOPE_REQUEST,
+                   Constants.SERVICE_RANKING + ":Integer=-700"
+
+           })
+public class LoggingFilter implements Filter {
+
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
+    @Override
+    public void doFilter(final ServletRequest request, final ServletResponse response,
+                         final FilterChain filterChain) throws IOException, ServletException {
+
+        final SlingHttpServletRequest slingRequest = (SlingHttpServletRequest) request;
+        logger.debug("request for {}, with selector {}", slingRequest
+                .getRequestPathInfo().getResourcePath(), slingRequest
+                .getRequestPathInfo().getSelectorString());
+
+        filterChain.doFilter(request, response);
+    }
+
+    @Override
+    public void init(FilterConfig filterConfig) {
+    }
+
+    @Override
+    public void destroy() {
+    }
+
+}
