+++ ui.apps/src/main/content/jcr_root/apps/training/aaaaa/SampleTests.js
