+++ ui.apps/src/main/content/jcr_root/apps/training/components/page-content/page-content.html
@@ -1,0 +1,7 @@
+<div>
+    <div>logo</div>
+    <div>
+        <div>toptoolbar</div>
+    </div>
+    <div data-sly-resource="${'topnav' @ resourceType='training/components/topnav'}"></div>
+</div>
