+++ ui.apps/src/main/content/jcr_root/apps/training/components/topnav/.content.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
+          jcr:primaryType="cq:Component"
+          jcr:title="Training Top Navigation Component"
+          jcr:description="Top Navigation for the Training Project"
+          sling:resourceSuperType="/training/templates/page-content/"
+          componentGroup="training"/>
